admin/admin/admin_modules/authentication/serilalizers.py
from rest_framework import serializers

from admin_modules.authentication.models import User


class UserSerializer(serializers.ModelSerializer):
    class Meta:
        model = User
        fields = "__all__"
